+++ src/main/java/me/pajic/affogato_core/mixson/MixsonHelper.java
@@ -2,20 +2,52 @@ package me.pajic.affogato_core.mixson;
 
 import com.google.gson.JsonElement;
 import me.pajic.affogato_core.Main;
+import net.minecraft.server.packs.resources.Resource;
 import net.ramixin.mixson.Mixson;
 import net.ramixin.mixson.MixsonCodecs;
 import net.ramixin.mixson.enums.ErrorPolicy;
 import net.ramixin.mixson.enums.Lifetime;
 import net.ramixin.mixson.util.Index;
 import net.ramixin.mixson.util.functions.Event;
+import net.ramixin.mixson.util.interfaces.MixsonCodec;
 
-import java.awt.image.BufferedImage;
+import java.io.BufferedReader;
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
 import java.util.Set;
 import java.util.function.Predicate;
 
 public class MixsonHelper {
 
     private static final ErrorPolicy ERROR_POLICY = Main.DEBUG ? ErrorPolicy.THROW : ErrorPolicy.LOG;
+
+    public static final MixsonCodec<String> TEXT = new MixsonCodec<>() {
+        public String extensionAndDot() {
+            return ".txt";
+        }
+
+        @Override
+        public String deserialize(Resource r) throws IOException {
+            String contents;
+            try (BufferedReader reader = r.openAsReader()) {
+                contents = reader.readAllAsString();
+            }
+            return contents;
+        }
+
+        @Override
+        public Resource serialize(Resource r, String s) {
+            return new Resource(r.source(), () -> new ByteArrayInputStream(s.getBytes()), r::metadata);
+        }
+
+        @Override
+        public ByteArrayOutputStream export(String s) throws IOException {
+            ByteArrayOutputStream baos = new ByteArrayOutputStream();
+            baos.write(s.getBytes());
+            return baos;
+        }
+    };
 
     public static void registerSingleJson(String eventName, String target, Event<JsonElement> event) {
         Mixson.registerEvent(
@@ -53,14 +85,14 @@ public class MixsonHelper {
         );
     }
 
-    public static void registerSingleTexture(String eventName, String target, Event<BufferedImage> event) {
+    public static void registerMultiText(String eventName, Predicate<Index> resourcePredicate, Event<String> event) {
         Mixson.registerEvent(
-                MixsonCodecs.PNG,
+                TEXT,
                 Mixson.DEFAULT_PRIORITY,
                 Lifetime.PERSISTENT,
                 ERROR_POLICY,
                 eventName,
-                index -> index.idEquals(new Index(target)),
+                resourcePredicate,
                 event
         );
     }